Refactoring and Corrections

+++ src/main/java/com/learn/NotificationService/controller/NotificationController.java
@@ -1,9 +1,6 @@
 package com.learn.NotificationService.controller;
 
-import com.learn.NotificationService.model.BlacklistRequest;
-import com.learn.NotificationService.model.BlacklistResponse;
-import com.learn.NotificationService.model.SmsRequest;
-import com.learn.NotificationService.model.SmsResponse;
+import com.learn.NotificationService.model.*;
 import com.learn.NotificationService.model.entity.SmsRequestDetails;
 import com.learn.NotificationService.service.NotificationService;
 import lombok.extern.slf4j.Slf4j;
@@ -12,6 +9,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.HttpClientErrorException;
 
 import javax.ws.rs.BadRequestException;
 import javax.ws.rs.Consumes;
@@ -24,7 +22,6 @@ import javax.ws.rs.Produces;
 @RequestMapping("/v1")
 public class NotificationController {
     private final NotificationService notificationService;
-
     public NotificationController(NotificationService notificationService) {
         this.notificationService = notificationService;
     }
@@ -36,16 +33,21 @@ public class NotificationController {
         SmsResponse smsResponse = new SmsResponse();
         HttpStatus httpStatus = HttpStatus.OK;
         try {
+            if(smsRequest.getPhoneNumber() == null || smsRequest.getMessage()== null)
+                throw new HttpClientErrorException(HttpStatus.BAD_REQUEST, "Phone number or message missing");
             log.info("sms request recieved : {}", smsRequest);
-            smsResponse = notificationService.sendAndSaveSms(smsRequest);
-        } catch (BadRequestException e) {
+            smsResponse.setData(notificationService.sendAndSaveSms(smsRequest));
+        } catch (BadRequestException |HttpClientErrorException e) {
+            SmsError smsError = new SmsError(String.valueOf(HttpStatus.BAD_REQUEST), e.getMessage());
+            smsResponse.setError(smsError);
             log.error("BadRequestException {}", ExceptionUtils.getStackTrace(e));
             httpStatus = HttpStatus.BAD_REQUEST;
         } catch (Exception e) {
+            SmsError smsError = new SmsError(String.valueOf(HttpStatus.INTERNAL_SERVER_ERROR), e.getMessage());
+            smsResponse.setError(smsError);
             log.error("Exception {}", ExceptionUtils.getStackTrace(e));
             httpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
         }
-        log.info("got response: {}", smsResponse.toString());
         return ResponseEntity.status(httpStatus).body(smsResponse);
 
     }
@@ -53,72 +55,92 @@ public class NotificationController {
     @Consumes("application/json")
     @Produces("application/json")
     @PostMapping("/blacklist")
-    public ResponseEntity<String> addNumbersToBlacklist(@RequestBody BlacklistRequest blacklistRequest) {
-        String response = "";
+    public ResponseEntity<BlacklistResponse> addNumbersToBlacklist(@RequestBody BlacklistRequest blacklistRequest) {
+        BlacklistResponse response = new BlacklistResponse();
         HttpStatus httpStatus = HttpStatus.OK;
         try {
+            if(blacklistRequest==null)
+                throw new HttpClientErrorException(HttpStatus.BAD_REQUEST, "Null Request");
             log.info("request recieved to blacklist numbers : {}", blacklistRequest);
             response = notificationService.addNumbersToBlacklist(blacklistRequest);
-        } catch (BadRequestException e) {
-            log.error("BadRequestException {}", e.getMessage());
+        } catch (BadRequestException | HttpClientErrorException e) {
+            log.error("BadRequestException {}", ExceptionUtils.getStackTrace(e));
             httpStatus = HttpStatus.BAD_REQUEST;
         } catch (Exception e) {
-            log.error("Exception {}", e.getMessage());
+            log.error("Exception {}", ExceptionUtils.getStackTrace(e));
             httpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
         }
         return ResponseEntity.status(httpStatus).body(response);
     }
     @Produces("application/json")
     @GetMapping("/blacklist")
-    public ResponseEntity<BlacklistResponse> getBlacklistedNumbers() {
-        BlacklistResponse blacklistResponse = new BlacklistResponse()
+    public ResponseEntity<BlacklistNumbersResponse> getBlacklistedNumbers() {
+        BlacklistNumbersResponse blacklistNumbersResponse = new BlacklistNumbersResponse()
 ;        HttpStatus httpStatus = HttpStatus.OK;
         try {
             log.info("getting blacklisted numbers ");
-            blacklistResponse = notificationService.getBlacklistedNumbers();
+            blacklistNumbersResponse = notificationService.getBlacklistedNumbers();
         } catch (BadRequestException e) {
-            log.error("BadRequestException {}", e.getMessage());
+            log.error("BadRequestException {}", ExceptionUtils.getStackTrace(e));
             httpStatus = HttpStatus.BAD_REQUEST;
         } catch (Exception e) {
-            log.error("Exception {}", e.getMessage());
+            log.error("Exception {}", ExceptionUtils.getStackTrace(e));
             httpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
         }
-        return ResponseEntity.status(httpStatus).body(blacklistResponse);
+        return ResponseEntity.status(httpStatus).body(blacklistNumbersResponse);
     }
     @Consumes("application/json")
     @Produces("application/json")
     @DeleteMapping("/blacklist")
-    public ResponseEntity<String> deleteNumbersFromBlacklist(@RequestBody BlacklistRequest blacklistRequest) {
-        String response = "";
+    public ResponseEntity<BlacklistResponse> deleteNumbersFromBlacklist(@RequestBody BlacklistRequest blacklistRequest) {
+        BlacklistResponse blacklistResponse = new BlacklistResponse();
         HttpStatus httpStatus = HttpStatus.OK;
         try {
+            if(blacklistRequest==null)
+                throw new HttpClientErrorException(HttpStatus.BAD_REQUEST, "Null Request");
             log.info("request recieved to delete numbers from the blacklist: {}", blacklistRequest);
-            response = notificationService.deleteNumbersFromBlacklist(blacklistRequest);
-        } catch (BadRequestException e) {
-            log.error("BadRequestException {}", e.getMessage());
+            blacklistResponse = notificationService.deleteNumbersFromBlacklist(blacklistRequest);
+        } catch (BadRequestException | HttpClientErrorException e) {
+            log.error("BadRequestException {}", ExceptionUtils.getStackTrace(e));
             httpStatus = HttpStatus.BAD_REQUEST;
         } catch (Exception e) {
-            log.error("Exception {}", e.getMessage());
+            log.error("Exception {}", ExceptionUtils.getStackTrace(e));
             httpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
         }
-        return ResponseEntity.status(httpStatus).body(response);
+        return ResponseEntity.status(httpStatus).body(blacklistResponse);
     }
 
-    //custom exception handling
+    @Consumes("application/json")
+    @Produces("application/json")
     @GetMapping("/sms")
-    public ResponseEntity<SmsRequestDetails> getSmsDetails(@RequestParam(value = "request_id") Integer orderId) {
-        SmsRequestDetails smsRequestDetails = new SmsRequestDetails();
+    public ResponseEntity<SmsDetailResponse> getSmsDetails(@RequestParam(value = "request_id") Integer orderId) {
+        SmsRequestDetails smsRequestDetails;
+        SmsDetailResponse smsDetailResponse = new SmsDetailResponse();
         HttpStatus httpStatus = HttpStatus.OK;
         try {
+            if(orderId==null)
+                throw new HttpClientErrorException(HttpStatus.BAD_REQUEST, "OrderId is null");
             log.info("request recieved for order details for Order ID : {}", orderId);
             smsRequestDetails = notificationService.getSmsDetails(orderId);
+            if(smsRequestDetails==null)
+                throw new HttpClientErrorException(HttpStatus.NOT_FOUND, "No order found for given id");
+            smsDetailResponse.setData(smsRequestDetails);
         } catch (BadRequestException e) {
-            log.error("BadRequestException {}", e.getMessage());
+            log.error("BadRequestException {}", ExceptionUtils.getStackTrace(e));
             httpStatus = HttpStatus.BAD_REQUEST;
+            SmsError smsError = new SmsError(String.valueOf(HttpStatus.BAD_REQUEST), e.getMessage());
+            smsDetailResponse.setError(smsError);
+        } catch(HttpClientErrorException e){
+            log.error("BadRequestException {}", ExceptionUtils.getStackTrace(e));
+            httpStatus = HttpStatus.NOT_FOUND;
+            SmsError smsError = new SmsError(String.valueOf(HttpStatus.NOT_FOUND), e.getMessage());
+            smsDetailResponse.setError(smsError);
         } catch (Exception e) {
             log.error("Exception {}", e.getMessage());
             httpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
+            SmsError smsError = new SmsError(String.valueOf(HttpStatus.INTERNAL_SERVER_ERROR), e.getMessage());
+            smsDetailResponse.setError(smsError);
         }
-        return ResponseEntity.status(httpStatus).body(smsRequestDetails);
+        return ResponseEntity.status(httpStatus).body(smsDetailResponse);
     }
 }
